requirements.txt
google-genai>=1.0.0
google-cloud-aiplatform[adk,agent_engines]>=1.95.0
google-cloud-storage>=2.14.0
google-cloud-bigquery>=3.20.0
google-adk>=1.0.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.1
tenacity>=8.2.3
pydantic>=2.6.0
typer>=0.12.0
rich>=13.7.0
